successful api call and added env var
